requirements.txt
google-genai
faiss-cpu
numpy
streamlit
